user_manager: sync timezones between host and guest
now you might be wondering why we're using setprop of container manager and not session manager even tho the session manager one is cleaner
well, i'm sorry to break it to you, i'm not a massive idiot. there is an actual reason behind this
IPlatform and all other services we implement might not be (and in fact, will not be in many of the attempts) fully up yet when we call them so Setprop of sm get stuck and block the mainloop
we want to call setprop as soon as we get that user 0 is ready because apps need to reflect this. so using container manager over lxc gives us much more consistent results

Signed-off-by: Bardia Moshiri <bardia@furilabs.com>